agents/travel_agent/requirements.txt
# Microsoft Agent Framework
agent-framework>=1.0.0b251108

# FastAPI for REST API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Azure SDK
azure-identity>=1.19.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
